Long -> Integer (#420)

+++ src/main/java/no/nav/veilarbaktivitet/aktivitet/AktivitetDataRowMapper.java
@@ -114,8 +114,8 @@ public class AktivitetDataRowMapper implements RowMapper<AktivitetData> {
 
     private static SokeAvtaleAktivitetData mapSokeAvtaleAktivitet(ResultSet rs) throws SQLException {
         return SokeAvtaleAktivitetData.builder()
-                .antallStillingerSokes(rs.getLong("antall_stillinger_sokes"))
-                .antallStillingerIUken(rs.getLong("antall_stillinger_i_uken"))
+                .antallStillingerSokes(rs.getInt("antall_stillinger_sokes"))
+                .antallStillingerIUken(rs.getInt("antall_stillinger_i_uken"))
                 .avtaleOppfolging(rs.getString("avtale_oppfolging"))
                 .build();
     }
